Improving handling of module initialization
Former-commit-id: 2e1842dce5afb06026d5e1c29fafa7e7c274bc88
Former-commit-id: b0868a4c1662446d0a2e4e363ca7d04ba7046a04
Former-commit-id: d514a307bc014627a62cb94e053ff282499178b8
Former-commit-id: 17c58b10bd5a5067b3794b7c24af277a79840871
Former-commit-id: cc8972fcd66d851851692e0f7dcafeb182d94645
Former-commit-id: e6367362a948a09b74c61bcd77997048dda1179f
Former-commit-id: 089918b063e1e9bbc4eadddfe76dbda5eacc5ea8
Former-commit-id: 1acfe447e7b3943ba64664bf786fd5ed1bbc18ec

+++ PowerShellConsole/ModuleInitialization/4 - General/Set Custom Prompt.ps1
@@ -1,3 +1,6 @@
+param([switch]$GetModuleInitStepRunLevel)
+if ($GetModuleInitStepRunLevel) { return 2 }
+
 if (Test-Path -Path Function:Global:prompt) {
     Remove-Item -Path Function:Global:prompt -Force
 }
